+++ src/main/java/br/com/rd/projetoVelhoLuxo/repository/contract/SubjectRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubjectRepository extends JpaRepository<Subject, Byte> {
+public interface SubjectRepository extends JpaRepository<Subject, Long> {
